backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
pydantic==2.6.1
grpcio==1.62.0
google-api-core==2.17.1
google-auth==2.28.1 